Folders crud

+++ resources/assets/js/src/medialib/components/Media.vue
@@ -224,7 +224,10 @@
                 this.$store.dispatch('editFolder', payload)
             },
             removeFolder(active) {
-                this.$store.dispatch('removeFolder', active)
+                let c = confirm("Are you sure?")
+                if (c === true) {
+                    this.$store.dispatch('removeFolder', active)
+                }
             }
         }
     }
